% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{block_spacer}
\alias{block_spacer}
\title{A vertical spacer block}
\usage{
block_spacer(height)
}
\arguments{
\item{height}{height in points}
}
\value{
A block object: a \code{list} with components \code{measure},
  \code{draw}, \code{keep} and \code{split} that the flow engine
  understands. When placed by \code{\link{flow_run}} or
  \code{\link{flow_add}} it consumes \code{height} points of vertical
  space without drawing anything.
}
\description{
A vertical spacer block
}
